add debug, and fix byobu configuration file copy

+++ provisioning/bootstrap.sh
@@ -10,13 +10,17 @@ debconf-set-selections <<< 'mysql-server-5.5 mysql-server/root_password_again pa
 apt-get install -qy nginx php5 php5-fpm php5-gd php5-mysql php5-cgi php5-cli php5-curl php5-xdebug vim git-core mysql-server-5.5 mysql-client curl byobu
 
 echo "General config stuff"
-mkdir /home/vagrant/.byobu
+if [ ! -d /home/vagrant/.byobu ]; then
+  mkdir /home/vagrant/.byobu
+fi
 cp /vagrant/provisioning/configs/.byobu/keybindings.tmux /home/vagrant/.byobu/
 cp /vagrant/provisioning/configs/.byobu/status /home/vagrant/.byobu/
 sudo chown -R vagrant:vagrant /home/vagrant/.byobu/
 
 echo "Moving nginx config files into place…"
-rm /etc/nginx/sites-enabled/default
+if [ -f /etc/nginx/sites-enabled/default ]; then
+  rm /etc/nginx/sites-enabled/default
+fi
 cp /vagrant/provisioning/nginx/default.conf /etc/nginx/sites-enabled/
 # cp /vagrant/provisioning/nginx/fastcgi_params.conf /etc/nginx/
 cp /vagrant/provisioning/nginx/dummy.* /etc/nginx/
@@ -28,6 +32,7 @@ cp /vagrant/provisioning/mysql/my.cnf /etc/mysql/my.cnf
 echo "Moving php config files into place…"
 mv /etc/php5/fpm/php.ini /etc/php5/fpm/php.ini.default
 cp /vagrant/provisioning/php/php.ini /etc/php5/fpm/php.ini
+sudo cp /vagrant/provisioning/php/20-xdebug.ini /etc/php5/fpm/conf.d/
 
 echo "Starting services…"
 service nginx restart
